+++ optimal/src/createPredicate.ts
@@ -1,6 +1,8 @@
 import { AnySchema, InferSchemaType, Predicate } from './types';
 
-export function createPredicate<T extends AnySchema>(schema: T): Predicate<InferSchemaType<T>> {
+export function createPredicate<S extends AnySchema, T = InferSchemaType<S>>(
+	schema: S,
+): Predicate<T> {
 	return (value) => {
 		try {
 			schema.validate(value);