+++ optimal/src/schemas/blueprint.ts
@@ -1,7 +1,7 @@
-import { AnySchema } from '../types';
+import { AnySchema, DefaultValue } from '../types';
 import { object } from './object';
 import { schema } from './schema';
 
-export function blueprint(defaultValue?: Record<string, AnySchema>) /* infer */ {
+export function blueprint(defaultValue?: DefaultValue<Record<string, AnySchema>>) /* infer */ {
 	return object(defaultValue).of(schema().notNullable());
 }