+++ optimal/src/schemas/custom.ts
@@ -1,6 +1,6 @@
 import { createSchema } from '../createSchema';
 import { commonCriteria } from '../criteria';
-import { CommonCriterias, CustomCallback, Schema } from '../types';
+import { CommonCriterias, CustomCallback, DefaultValue, Schema } from '../types';
 
 export interface CustomSchema<T> extends Schema<T>, CommonCriterias<CustomSchema<T>> {
 	never: () => CustomSchema<never>;
@@ -8,7 +8,7 @@ export interface CustomSchema<T> extends Schema<T>, CommonCriterias<CustomSchema
 	nullable: () => CustomSchema<T | null>;
 }
 
-export function custom<T>(callback: CustomCallback<T>, defaultValue?: T) {
+export function custom<T>(callback: CustomCallback<T>, defaultValue?: DefaultValue<T>) {
 	return createSchema<CustomSchema<T>>({
 		criteria: { ...commonCriteria },
 		defaultValue,